Move file handling across to a new API.  (#137)  Sorry if this landing screws anyone up.
main-gtk is untouched, because I don't want to conflict with the work that's been going on outside the repository on main-gtk.

Showing scores from the commandline seems broken right now, but other than that, this works fine on at least Linux and Windows.

Tests included in a hacked-together file; may prove useful to e.g. making sure it works on OS X.



git-svn-id: svn://rephial.org/angband/trunk@469 98ddb578-d4ae-11db-a93f-5346e46f259c