Use a new algorithm for how much gold is in each gold drop.  It replaces to the previous and somewhat insane system such that gold increases linearly with level.
The graph at http://rephial.org/research/avg_gold_drop.pdf shows the effect of the change; in short, gold amounts up to dungeon level 40 are mostly the same, but vastly reduced after that.  Additionally, there is less variance in the amount of gold (it is now 50%), such that most dungeon levels now have a higher minimum gold drop but a lower maximum.


git-svn-id: svn://rephial.org/angband/trunk@889 98ddb578-d4ae-11db-a93f-5346e46f259c