Preserve menu cursor when updating menu contents (fixes #1236)
This only really affects stores, where the store contents needs to be updated
after a transaction, but the current cursor position needs to be preserved.
If the current position is off the end, drop back to the last position.

Menus are initialised with the cursor on the first element (row = 0).

Signed-off-by: Peter Denison <angband@marshadder.org>