Tweaks to Stealth to make it less dominating.
This patch mostly removes a few points of stealth, basically every
reduction is for 1 point, and many places have not been touched. But
in concert the change will be felt. On average a starting character
will have 2 stealth less, and a character that is using three randomly
chosen stealth items will also on average have 2 less stealth from eq.

The only change that might be seen as radical is that I have changed the
Rogue class to start at 3 stealth, and increase by 1 / 10 lvls. When all
changes are accounted for this should mean that by the endgame a Rogue
trying to get stealthy should be just a bit better at it than he would be
today, while all other classes would be worse. Thus making the Rogue
class stand out more.

I gave the Elf race a boost in CON, from -2 to -1. I felt I needed to
reduce their stealth, but did not want to weaken them further. They
have long needed a boost in any case.

The changes to monster.txt are all to 'alertness'. I increased many values
in the early dlevels to compensate for reduced @ stealth. So, for the early
levels (10 first dlvls) characters should be, in general, about as
'stealthy' as before, relative to the monsters. In the later dlevels
some values have been lowered, and in general some values have been
modified for thematic reasons.

If 'cherry picking' from this commit, please note that this is meant as
a package. I feel it is balanced as a whole, and removing some parts means
that the changes would need to be balanced again. I also feel that since
this is more of a tweak than a change, there is enough time to get this
into 3.3.