Enable solid walls for GCU.
This commit moves from using unicode for semi-solid walls
to using the A_REVERSE attribute to get solid walls. This
is a bit more reliable and also nicer-looking.

There is one gotcha, which is that we use the high "color
bit" to determine when to use A_REVERSE. This lowers the
maximum number of possible colors from 255 to 127 (right
now we're using less than 40 I think).