Fallback to AC_CHECK_LIB if pkg-config check fails
Some systems (like OpenBSD) do not ship pkg-config/ncurses5-config information
for ncurses, since it's part of their base system, and not necessary. In that
case, fallback to checking for a function (`initscr`) in `libncursesw`.

Additionally, define `_XOPEN_SOURCE_EXTENDED`, which ensures that angband
picks up the right definitions of `mvwaddnwstr`.