Fix previous commit which removed curses in object_delete()
In the long term, object_delete()ing the object itself and its known
aspect separately seems likely to lead to memory leaks.  There are
also likely to be places where curses aren't freed since this is not
done in object_delete() precisely because the base object & its known
copy are deleted separately.