Introduce a new 'source' type for projection and effect sources
This is the first stage of a refactoring to remove mon_current and
trap_current.  project() and its sub-functions take 'struct source'
which is a type saying which trap, monster or player was the origin
of the effect.  This removes the hack where 'who' meant a monster
if it was above zero and player if less than zero, which also
didn't really work with traps as an effect source.