Fix "OBJECT hits MONSTER" message for throw and fire commands
Make make_ranged_shot() and make_ranged_throw() copy "hits" to hit_verb by using the allocated buffer size (20) instead of sizeof(hit_verb) as the length of the buffer. In a 32-bit environment, my_strcpy only copies fist three letters (the 4th element is the terminator '\0') as the value of sizeof(hit_verb) is 4. In the game, the bug made the message "OBJECT hit MONSTER" instead of "OBJECT hits MONSTER".