Review my changes to the 16x16 font and commit only a few strict, conservative improvements
Namely: comma made more distinct from dot, semicolon from colon,
backquote from quote and braces from parens.
Tilde centered so that it's on the same height as minus, etc.
The hash sign made square and so centered wrt other punctuation
(this one change is disputable, because in the effect, hash
is less centered wrt the bounding box, but this is just a deficiency
of this font and all other punctuation and middle dots suffer
from being too high relative to the bounding box in the same way).