Ensure press is on a item list before showing the context menu
If Player is in a store/home and clicks the header line (line 3,
where "Store/Home Inventry" is displayed) and the store/home is
empty (hence oid is -1 because cursor is not on a list item),
the game still thinks press is on a list item and tries to
show the store item context menu for non-existant item and make
an illegal memory access.

Checking if clicked location is actually on a list item (if y is
more than or equal to 4) ensures this does not happen.