Recalculate player equipment bonuses after digging.
When digging, the game switches you automatically to your best digger and recalculates your equipment bonuses.  After digging is complete it switches your weapon back, but doesn't recalculate equipment bonuses, so stat boosts from your weapon will not be applied until the next event that triggers a recalculation (such as equiping or taking off a piece of equipment).

This commit adds a recalculation after reequiping so that @s abilities will be accurate immediately.