Modify text when player is aware of a monster via detection or telepathy
commit 496fc1c51d5276100d924c5a64ca1585f676c205 changed the "You see" message
when examining a square to "You recall" if the square wasn't visible in order
to avoid "seeing" while blind.

This has the unfortunate side effect of using "You recall" when detecting or using
telepathy.

Detect whether a visible monster is on the square that wouldn't have otherwise
been seen and infer that it was viewable via detection or telepathy.  This
unfortunately doesn't handle non-monsters viewable this way (such as doors via
"Read Minds" or shop entrances), but it clears up the common case of monsters