Change default enter behavior in stores
This proposal changes what happens after a player presses the enter key within a store from the event being dropped to purchasing the item the cursor is under.

I don't know whether keeping the EVT_SELECT branch separate is cleaner than putting it as a parallel case with the EVT_KBRD handler.

The justification for this change is twofold. First, the player can move the cursor in the store, but when the player decides to do anything (examine or purchase), they have to reselect the item. I think enter defaulting to purchasing is natural. Secondly, this helps playing the game on platforms without hardware keyboards (namely consoles and phones) but doesn't change any of the underlying behavior for PC players.