tests: run unit tests against in-tree data files
Before this change, unit tests ran against the production data files on disk
(if the game was installed in $PREFIX) and otherwise just errored out. That
made the build non-hermetic and in particular made the test suite not work
in CI environments.

This change has the unit tests always look in $SRCROOT/lib for their data
files instead.

It also adds a "check" target, as an alias to "tests", since "make check"
is basically standard and some package managers (like apk) default to
running it.