Apply shape's attributes before applying the vulnerabilities from all sources
That lets elemental resistances stack correctly if the shape provides a vulnerability or a resistance that's not a immunity and the other thing affecting that resistance include both a resistance and a vulnerability.  Does not impact the shapes currently in Vanilla.  Pukel-man is the only one providing resistances:  one is an immunity and the other is for an element that does not have vulnerabilities.