Character sheet: change rendering of numeric modifiers
When directly rendering the number (unused in Vanilla) use the combined value and auxiliary value and allow coloring based on a tuple of the sign (0, +, -) of that combined value and the sign of the auxiliary value.  When changing the number to a symbol, use a tuple of the sign of the combined value and auxiliary value and the sign of the auxiliary value to determine the symbol.  With the default settings, show a "!" or "=" when the timed effect is not zero and has the same sign as the combined timed and non-timed effect (previously would only show those if the timed effect was not zero and the non-timed effect was zero).  When the timed effect is not zero and regardless of the combined value, change the color of the symbol ('G' if the timed effect is positive; 'o' if the timed effect is negative).  Allows resolving something noted in https://github.com/NickMcConnell/FAangband/issues/392 : a character that gets infravision from the character's race did not see the effect on the second character sheet from a potion of infravision.  Changes how the palettes are used for the NUMERIC_AS_SIGN_RENDERER_WITH_COMBINED_AUX and NUMBER_RENDERER_WITH_COMBINED_AUX renderers so if those have been customized by changes to ui_entry_renderers.txt or to preference files, those customizations will have to be adapted to these changes.