Extend mean() and variance()
Allow the caller to request the fractional part of the result.  If the fractional part is not requested, the returned value is rounded to the nearest integer rather than the prior behavior of rounding towards zero.  Avoid overflow in the internal calculations.  For variance() if the result would be greater than INT_MAX, return INT_MAX.  For variance(), give the option to normalize by size minus one rather than size.  Also give an option for variance() to scale the result by one over size to correspond to the expected variance of a mean estimate.

Use the new forms of mean() and variance() in the object damage information test cases.  Avoids the accumulation of roundoff error that the previous version of those test cases had.  Allows increasing the number of hit simulations in those tests to 1000000 or more without encountering test failures.