Allow an object curse to affect an object's weight
Idea stolen from will_asher's RubberBand.
These changes only provide the infrastructure for curses that affect weights; no new curses have been defined in curse.txt.
To get the weight, including the effect of curses, of an object, o, use object_weight_one(o).  It is declared in obj-util.h.
Both additive and multiplicative adjustments to the weight are allowed (but not in the same curse).  The latter use a new object flag, OF_MULTIPLY_WEIGHT.  In Vanilla, adding that flag does not change the size of the flags field, but that may not hold true for variants.
The object power calculations for randarts have some accounting for changes to weight:  ac_power() uses object_weight_one() when computing the AC per unit weight, weight-affecting curses get special treatment in curse_power(), and there is a new term in the power, nonstandard_weight_power(), to account for differences in the weight from the standard value.  Those likely need tuning.