Allow player pathfinding to go through closed doors, impassable rubble, traps
Resolves https://github.com/angband/angband/issues/5867 .
When running along the computed path, only open a door or clear impassable rubble if all the neighbors of the door or rubble are known.  Otherwise, stop before the door and rubble and let the player decide what to do.  For known visible traps, always stop unless the player is trapsafe.
When computing the path to a previously mapped location, prefer that the path only go through previously mapped squares.  if no such path is found, then allow going through squares that have not been mapped before.
When computing a path, prefer that the path does not include a known visible trap.  If no such path is found, then allow going through known visible traps.
For tests of known terrain when computing the path, use the player's memory of the cave so information is not leaked if the terrain has changed since the player saw it.  Similarly, when running along the computed path, tests of known terrain and objects use the player's memory so information is not leaked if pathfinding while blind or without a light source.
Move computing the chance of unlocking a locked door to its own function, calc_unlocking_chance() declared in player-calcs.h, so it can be shared by do_cmd_open_aux() and the pathfinding calculations.