Plug source of occasional test failure in tests/game/basic
The cause was moving away from the stairs and then have a monster move to the stairs before being able to move back.  Also, call on_new_level() after most reloads in that test so player->state and player->known_state are filled as they would be when play_game() is called.