Rename square_isnotknown(); make it and square_isknown() useful for player->cave
The new name for square_isnotknown() is square_ismemorybad().  In the case where it is passed a chunk that is neither cave nor player->cave, it will return true (no player memory).  That makes it equivalent to !square_isknown() and a check on whether the terrain in player->cave and cave is the same.  Resolves https://github.com/angband/angband/issues/5879 .