Refactor player-path
For calculating multiple paths from the same starting point, make the distance array calculation public and have it return its results rather than storing them in a static array.  In the internals of that calculation, replace the point set with a queue and explicitly have the edges of the cave as unreachable to simplify bounds checks.

With that distance calculation available, have the nearest staircase/nearest unexplored calculations in do_cmd_navigate_down(), do_cmd_navigate_up(), and do_cmd_explore() use the expected number of turns to navigate the path rather than distance().  Move the search for the nearest to path_nearest_known() and path_nearest_unknown() (declared in player-path.h; implemented in player-path.c).

Change the interface of find_path() so it returns the steps in the path rather than storing them in a static array.  Change struct player_upkeep so all the state for running with pathfinding is there.  Removes the running_with_pathfind element of that structure; test for a non-NULL steps field in that structure instead.

Change the internals of find_path() to use an A* algorithm.  Add a priority queue type in z-queue.{c,h} to use in that algorithm.

To compare the performance before and after the change, did this experiment: generate 500 random levels at a depth of 1000 feet; for each level, memorized all grids in the level that did not seem like a wall or which touched at least one grid that did not seem like a wall; then randomly selected twenty destinations in the level that were known, passable, and different from the player's starting location and for each of those destinations timed the path computations using the previous version of find_path(), the modified version of find_path(), and the combination of prepare_pfdistances(), pfdistances_to_path(), and release_pfdistances().  The results were collected on a Linux virtual machine (Debian 11 running on macOS 12.7.2 host via Parallels 18) with a 2.7 GHz Dual-Core Intel Core i5 as the processor.

Release build (CFLAGS=-O2 ; --enable-release)
                        count   mean time (seconds)     std. dev. of mean (s)
old find_path()
    successful paths     9395        0.0012599                0.0000031
    failed paths          605        0.0032737                0.0000332
new find_path()
    successful paths     9395        0.0004527                0.0000039
    failed paths          605        0.0006668                0.0000543
pf_distances_to_path()
    successful paths     9395        0.0008525                0.0000029
    failed paths          605        0.0020103                0.0000318

Non-release build (CFLAGS=-O0 ; --disable-release)
                        count   mean time (seconds)     std. dev. of mean (s)
old find_path()
    successful paths    9277         0.0014400                0.0000041
    failed paths         723         0.0036938                0.0000369
new find_path()
    successful paths    9271         0.0006211                0.0000055
    failed paths         729         0.0012218                0.0000887
pf_distances_to_path()
    successful paths    9271         0.0010382                0.0000038
    failed paths         729         0.0027164                0.0000468