Improve curses affecting +hit, +dam, or +ac
If the curse affects +hit or +dam and is on the equipped weapon, equipped launcher, a thrown weapon, or launched ammunition, include the curse's affects in attack calculations and simulations of those attack calculations for damage information.  For such curses on the equipped weapon or equipped launcher, include the curse's affects in the combat information on the first page of the character sheet.  Resolves https://github.com/angband/angband/issues/5912 .

Show the affects of the curse on the +hit, +dam, and +ac values shown in object descriptions.  Resolves https://github.com/angband/angband/issues/5857 .