Mark commands used internally by pathfinding as background commands
Background commands now come in two flavors:  those that can trigger bloodlust and those that do not.  Generally gives better behavior than 4.2.5 when repeating the last command after an interrupted pathfinding command:  repeats the pathfinding command rather than the last run used to take a step along the path.  Fixes a regression introduced by 4d76112f25e5287b9c2ebfdaef80c72d13928900 : when compiled with -fsanitize=address and -fsanitize=undefined, repeating the last command after an interrupted pathfind when no normal run had been used in the session would trigger a crash in run_test() (in that case, the last command is run with a direction of zero and prev_run_dir is zero).

Has the command queue manage the lifetime of string arguments to commands.  Resolves a use after free if the command to inscribe an object is followed immediately by the command to repeat the last command.