SDL2: store separate layouts for fullscreen/not fullscreen modes
Resolves https://github.com/angband/angband/issues/5840 .  Resolves https://github.com/angband/angband/issues/5841 .

Adds entries to sdl2init.txt ("window-full-rect-fs" and "subwindow-full-rect-fs") which are not backwards compatible with previous versions.  To go back when sdl2init.txt has "window-fullscreen:true", delete the "window-full-rect:" and "subwindow-full-rect:" lines and then change "window-full-rect-fs" to "window-full-fect" and "subwindow-full-rect-fs" to "subwindow-full-rect".  To go back when sdl2init.txt had "window-fullscreen:false", delete the "window-full-rect-fs" and "subwindow-full-rect-fs" lines.

Does not allow the fullscreen and not fullscreen modes to have different configurations for the font, subwindow visibility, and subwindow transparency.