Clean up configure a little
Replace instances of '==' with '=' when used with the test command line utility. '==' is not documented and might lead to portability problems.  Put the logic for the spoiler, statistics, and test front ends before the logic for the Windows front end so the resetting of MAINFILES by the Windows front end has its full intended effect.