SDL2: provide ways to summarize details about the SDL2 installation
The intended use is for diagnosing problems with the SDL2 front end.  The -v suboption, "angband -msdl2 -- -v", will cause details about the SDL2 installation to be pushed through SDL_Log() so those details will be printed on standard error for most platforms (Windows and Android are the known exceptions).  The "SDL Details..." entry in the first menu will dump the details to a simple dialog.  That menu entry for the main window will provide more information (library versions, platform, current video driver, list of connected displays, and audio information).  For other windows, only the details specific to that window are displayed.  For problems that cause the application to exit in init_systems(), only the -v suboption will be useful, and it will only provide limited information in that case:  the version of the SDL library.