macOS: allow dead key events to be passed to the game's core
A dead key is a key combination that does not have a Unicode equivalent with the current keyboard mapping.  Potentially allows for more key combinations to be used as keymap triggers:  on macOS 12.7.6 with the US keyboard mapping, Option+E is passed to the game's core as {A}e rather than being ignored.