ui-game: don't attempt to open empty paths
This mimics a check which we have a few lines below: if the path is an
empty string, do not attempt to open it. This fixes an issue on the Wii,
where for some reason opening an empty path is successful, but it seems
a reasonable change to have on all platforms.