Autorepeat using a wand/staff/rod/activatable item on failure
Autorepetition only happens if there's a chance that the repetition will succeed.  The autorepetition is interrupted by anything that disturbs the player or by a player interrupt.

Recommended as part of the response to 4.2.3 change to the device failure calculation, for instance see https://angband.live/forums/forum/angband/vanilla/10352-angband-4-2-4?p=225500#post225500 .