Persistent levels:  don't allow reset of recall depth
Since the player has the choice of recall depth with persistent levels, resetting the recall and maximum depth only artficially limits which levels can be the destination for the recall.  In https://github.com/angband/angband/issues/6190 , very likely was the reason why the player could not recall to a previously visited level.

Also address two corner cases when getting the recall depth with persistent levels:  if the character has not visited the dungeon or is playing with forced descent on, there's no real choice for the destination so do not prompt for it.