Allow the pit statistics to loop over depth and write results to a file
The number of simulations is now the number of simulations to perform at each depth.  The results written to the message window, when looping over multiple depths, are the sum of the histograms at each depth.  The histograms at each depth and the sum over depth of those histgrams are written to the file.  Shorten the messages written to the message window:  what was "Type: Animals, Number: 310." is now "Type Animals: 310.".