gcu: allow compiling with mouse and resizing support for PDCurses
Because main-gcu.c uses the interface from NCurses for handling mouse events, the preprocessor macro, PDC_NCMOUSE, will have to be set when compiling main-gcu.c with PDCurses in order to have mouse support.  Further work may be necessary to actually have functioning mouse events and resizing as testing with PDCurses built on top of SDL2 running in a Debian 12 virtual machine hosted by Parallels on a Mac laptop (no mouse, only a trackpad) indicated these problems:  the game received separate button press and release events but they were not close enough in time for PDCurses to send a click event which is what main-gcu.c expects and making the window bigger did not trigger the game to fill in the extra space.