When called from get_aim_dir(), do not allow targeting to start pathfinding
Do allow that when called to implement the look and targeting commands or to implement looking in the context menus.  Changes the prototype for target_display_help().  As that function is only used in ui-target.c, limit the visibility of that function and the related HELP_HEIGHT to ui-target.c.