SDL2: be more thorough when cleaning up
Pair the call to init_graphics_modes() in init_sdl2() with a call to close_graphics_modes() in quit_hook().  When exiting because of an SDL_QUIT event, closing the main window, or by the Quit menu entry, mimic main.c's cleanup.

If there's a game in progress, there will still be leaked memory when exiting by SDL_QUIT/closing the window/Quit menu entry because close_game() is not called.  Best to take care of that when resolving https://github.com/angband/angband/issues/6459 .