Add column in monster knowledge menu to indicate if fully known
To keep within an 80 column limit and avoid cutting off a long name for a type of monster, that column's title is shortened to "Full".  Add a category, "Not Fully Known", which will display any types of monsters that have been seen but are not fully known.  If all monsters that have been seen are fully known or the player has copied over monster.txt to lore.txt, that category will be empty and will not be displayed.  Resolves https://github.com/angband/angband/issues/6361 .