Change introduction to the description of an item's effect
If the effect requires aiming, now preface that description with "It requires a target."  If the effect is unknown, assume it does not require aiming except for wands, rods, and activatable equipment.  For unknown wands, assume aiming is required.  For unknown rods and activatable equipment, be ambiguous and say "It may require a target."  For wand, staffs, rods, and activatable equipment use "used" as the verb since their effects can be accessed with the keyset's generic "use" command.  Resolves https://github.com/angband/angband/issues/6437 .